Rename server to oauth2
